refactor: rename directory-presence -> crawler-presence (review feedback)
Per @TaekeK on the PR: the lookup checks what the crawler has loaded, not the live
directory, so 'crawler presence' is the accurate name.
- directoryPresence.js -> crawlerPresence.js (+ test)
- lookupDirectoryPresence -> lookupCrawlerPresence; normalizeDirectoryQuery -> normalizePresenceQuery
- POST /ingest/principals-in-directory -> /ingest/principals-presence
- directoryAvailable -> crawlerDataAvailable (crawler call + comments updated)
